refactor: trim deployer-alpha
