+++ deployer-alpha/Cargo.toml
@@ -10,12 +10,10 @@ anyhow = { workspace = true }
 async-trait = { workspace = true }
 axum = { workspace = true, features = ["headers", "json", "query", "ws"] }
 bytes = { workspace = true }
-cargo = { workspace = true }
 cargo_metadata = { workspace = true }
 chrono = { workspace = true }
 clap = { workspace = true }
 crossbeam-channel = { workspace = true }
-flate2 = { workspace = true }
 fqdn = { workspace = true }
 futures = { workspace = true }
 home = { workspace = true }
@@ -38,7 +36,6 @@ sqlx = { workspace = true, features = [
   "uuid",
 ] }
 strum = { workspace = true }
-tar = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true, features = ["fs", "process"] }
 toml = { workspace = true }
@@ -69,6 +66,5 @@ features = ["builder"]
 
 [dev-dependencies]
 ctor = { workspace = true }
-hex = "0.4.3"
 rand = { workspace = true }
 tempfile = { workspace = true }
